+++ CHANGELOG.md
@@ -1,7 +1,12 @@
-## 2.1.1
+## 2.2.0
 
 * Support Animations [SlideTransition, FadeTransition, ScaleTransition, RotationTransition,     SizeTransition]
 * Support new Routes [DialogRoute, RawDialogRoute, PageRouteBuilder, ModalBottomSheetRoute]
 * Support passing T to Routes and return values
 * Support global access to [Anything.of(context)] using class [Gof]
 * Support show [dialog, bottomSheet, snackBar, and materialBanner] using class [GoMessenger]
+
+* Version 2.2.0
+* Editing three animation routes classes [GoRotation, GoScale, GoSize]
+* We added an alignment parameter to [GoRotation, GoScale]
+* and added another two [SizeDirection] in GoSize (right & left)